admin detect

+++ controller/admin.js
@@ -3,8 +3,8 @@ import Admin from "../model/admin.js"
 
 export const createAdmin = async (req, res, next) => {
     try {
-        const { username, password, role, ...adminInfo } = req.body
-        const account = await Account.createAccount(username, password, role)
+        const { username, password, ...adminInfo } = req.body
+        const account = await Account.createAccount(username, password, "admin")
         try {
             const admin = await Admin.create({
                 
@@ -23,7 +23,7 @@ export const createAdmin = async (req, res, next) => {
 
 export const getAdmin = async (req, res, next) => {
     try {
-        const admins = await Admin.find({})
+        const admins = await Admin.find({role: "admin"})
         return res.json(admins)
     } catch (error) {
         next(error)
@@ -33,6 +33,7 @@ export const getAdmin = async (req, res, next) => {
 export const getAdminById = async (req, res, next) => {
     try {
         const admin = await Admin.findById(req.params.id)
+        if (admin.role != "admin") return next(createError(400, "Admin not found"))
         if (!admin) return next(createError(400, "Admin not found"))
 
         return res.json(admin)
@@ -44,6 +45,7 @@ export const getAdminById = async (req, res, next) => {
 export const deleteAdminById = async (req, res, next) => {
     try {
         const admin = await Admin.findByIdAndDelete(req.params.id)
+        if (admin.role != "admin") return next(createError(400, "Admin not found"))
         if (!admin) return next(createError(400, "Admin not found"))
 
         const account = await Account.findByIdAndDelete(admin.accountId)
@@ -58,6 +60,7 @@ export const deleteAdminById = async (req, res, next) => {
 export const updateAdminById = async (req, res, next) => {
     try {
         const admin = await Admin.findByIdAndUpdate(req.params.id, req.body, { new: true })
+        if (admin.role != "admin") return next(createError(400, "Admin not found"))
         if (!admin) return next(createError(400, "Admin not found"))
         
         return res.json({ msg: `Update Admin successfully`,
